+++ src/ui/MessageMenu/index.tsx
@@ -2,3 +2,4 @@ import './index.scss';
 
 export { MessageMenu, type MessageMenuProps } from './MessageMenu';
 export { MessageMenuProvider, useMessageMenuContext, type MessageMenuContextProps } from './MessageMenuProvider';
+export { MenuItem, BottomSheetMenuItem, type MenuItemProps } from './menuItems/BasicItems';
